+++ ct-app/tools/hopr_api_helper.py
@@ -294,12 +294,12 @@ class HoprdAPIHelper:
             if source_peer_id not in peerid_address_aggbalance_links:
                 peerid_address_aggbalance_links[source_peer_id] = {
                     "source_node_address": source_address,
-                    "aggregated_balance": balance,
+                    "channels_balance": balance,
                 }
 
             else:
                 peerid_address_aggbalance_links[source_peer_id][
-                    "aggregated_balance"
+                    "channels_balance"
                 ] += balance
 
         return peerid_address_aggbalance_links
